moved helper read methods to mod read
