+++ serde_amqp/src/lazy.rs
@@ -2,87 +2,7 @@
 
 use bytes::Bytes;
 
-use crate::{format::Category, format_code::EncodingCodes, read::Read, Error};
-
-fn read_fixed_bytes<'de, R>(mut reader: R, width: usize) -> Result<(Vec<u8>, R), Error>
-where
-    R: Read<'de>,
-{
-    let bytes = reader.read_bytes(width + 1)?;
-    Ok((bytes, reader))
-}
-
-fn read_encoded_len<'de, R>(mut reader: R, width: usize) -> Result<(usize, R), Error>
-where
-    R: Read<'de>,
-{
-    let len_bytes = reader
-        .peek_bytes(width + 1)?
-        .ok_or_else(|| Error::unexpected_eof("parse LazyValue"))?;
-    match width {
-        1 => {
-            let len = len_bytes[1] as usize;
-            Ok((len, reader))
-        }
-        4 => {
-            let mut len_bytes_ = [0; 4];
-            len_bytes_.copy_from_slice(&len_bytes[1..]);
-            let len = u32::from_be_bytes(len_bytes_) as usize;
-            Ok((len, reader))
-        }
-        _ => unreachable!(),
-    }
-}
-
-fn read_encoded_len_bytes<'de, R>(reader: R, width: usize) -> Result<(Vec<u8>, R), Error>
-where
-    R: Read<'de>,
-{
-    let (len, mut reader) = read_encoded_len(reader, width)?;
-    let bytes = reader.read_bytes(1 + width + len)?;
-    Ok((bytes, reader))
-}
-
-fn read_primitive_or_else<'de, R, F>(mut reader: R, f: F) -> Result<(Vec<u8>, R), Error>
-where
-    R: Read<'de>,
-    F: FnOnce(R) -> Result<(Vec<u8>, R), Error>,
-{
-    let code: EncodingCodes = reader
-        .peek()
-        .ok_or(Error::unexpected_eof("parse LazyValue"))
-        .and_then(|code| code.try_into())?;
-
-    let (bytes, reader) = match Category::try_from(code) {
-        Ok(Category::Fixed(width)) => read_fixed_bytes(reader, width)?,
-        Ok(Category::Variable(width)) => read_encoded_len_bytes(reader, width)?,
-        Ok(Category::Compound(width)) => read_encoded_len_bytes(reader, width)?,
-        Ok(Category::Array(width)) => read_encoded_len_bytes(reader, width)?,
-        Err(_is_described) => f(reader)?,
-    };
-
-    Ok((bytes, reader))
-}
-
-fn read_described<'de, R>(mut reader: R) -> Result<(Vec<u8>, R), Error>
-where
-    R: Read<'de>,
-{
-    // Read 0x00
-    let mut bytes = reader.read_bytes(1)?;
-
-    // Read the descriptor
-    let (mut descriptor_bytes, reader) =
-        read_primitive_or_else(reader, |_| Err(Error::InvalidFormatCode))?;
-    bytes.append(&mut descriptor_bytes);
-
-    // Read the value
-    let (mut value_bytes, reader) =
-        read_primitive_or_else(reader, |_| Err(Error::InvalidFormatCode))?;
-    bytes.append(&mut value_bytes);
-
-    Ok((bytes, reader))
-}
+use crate::{read::{Read, read_primitive_bytes_or_else, read_described_bytes}, Error};
 
 /// A lazy value
 /// 
@@ -97,7 +17,7 @@ impl LazyValue {
     where
         R: Read<'de>,
     {
-        let (bytes, _) = read_primitive_or_else(reader, read_described)?;
+        let (bytes, _) = read_primitive_bytes_or_else(reader, read_described_bytes)?;
         Ok(LazyValue(Bytes::from(bytes)))
     }
 
@@ -131,7 +51,7 @@ impl From<LazyValue> for Vec<u8> {
 
 #[cfg(test)]
 mod tests {
-    use crate::{described::Described, descriptor::Descriptor, primitives::Array, to_vec};
+    use crate::{described::Described, descriptor::Descriptor, format_code::EncodingCodes, primitives::Array, to_vec};
 
     use super::*;
